refactor(types): reorganize clipboard and notification types for improved structure and clarity
