+++ src/types/clipboard.ts
