+++ src/types/index.ts
@@ -1,15 +1,19 @@
 export type { ActionResult } from "./actionResult";
 export type { AppResult } from "./appResult";
+export type { AppSettings } from "./AppSettings";
 export type { CategoryResults } from "./categoryResults";
-export type { SearchResultItem } from "./searchResultItem";
-export type { SearchResults } from "./searchResults";
-export type { ResultCategory } from "./resultCategory";
+export { ClipboardItemType, type ClipboardHistoryItem } from "./clipboard";
 export type {
   Extension,
   ExtensionManifest,
   ExtensionCommand,
   ExtensionResult,
+  SearchProvider,
 } from "./extension";
-export type { ClipboardHistoryItem } from "./clipboardHistoryItem";
-export { ClipboardItemType } from "./clipboardHistoryItem";
-export type { AppSettings } from "./AppSettings";
+export type {
+  NotificationActionType,
+  NotificationChannel,
+  NotificationOptions,
+} from "./Notification";
+export type { ResultCategory } from "./resultCategory";
+export type { SearchResults, SearchResultItem } from "./searchResults";
